+++ Lib/v3d/Invoke.h
@@ -1,6 +1,5 @@
 #ifndef _VC6_INVOKE_H
 #define _VC6_INVOKE_H
-//#include "Target/SharedArray.h"
 #include "SharedArray.h"
 #include "Common/Seq.h"
 
@@ -10,7 +9,7 @@ namespace v3d {
 
 void invoke(
   int numQPUs,
-  SharedArray<uint32_t> &codeMem,
+  SharedArray<uint64_t> &codeMem,
   int qpuCodeMemOffset,
   Seq<int32_t>* params);
